#34 Moved methods from Controller
