+++ src/test/kotlin/org/oso/core/controller/EmergencyControllerTest.kt
@@ -9,10 +9,12 @@ import org.mockito.Mockito
 import org.oso.any
 import org.oso.core.controllers.EmergencyController
 import org.oso.core.dtos.EmergencyDto
+import org.oso.core.dtos.EmergencyAcceptedDto
 import org.oso.core.entities.Emergency
 import org.oso.core.entities.EmergencyPriority
 import org.oso.core.entities.HelpRequester
 import org.oso.core.services.EmergencyService
+import org.oso.core.services.HelpProviderService
 import org.oso.core.services.HelpRequesterService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs
@@ -35,6 +37,9 @@ class EmergencyControllerTest {
 
     @MockBean
     lateinit var emergencyService: EmergencyService
+
+    @MockBean
+    lateinit var helpProviderService: HelpProviderService
 
     @MockBean
     lateinit var helpRequesterService: HelpRequesterService
@@ -111,5 +116,41 @@ class EmergencyControllerTest {
 //
 //        // TODO test fails if helprequester is not given, as the helprequester-id will currently be defaulted to 0 which leads to HTTP 404 instead of 400
 //        testFailOnMissingParameters(mockMvc, "/${EmergencyController.PATH_EMERGENCY}/${EmergencyController.PATH_EMIT}", parameters)
+    }
+
+    @Test
+    fun testAcceptEmergency() {
+        val dto = EmergencyAcceptedDto(
+                emergencyId = "25",
+                helpProviderId = "38",
+                helpRequesterId = "46"
+        )
+
+        Mockito.doNothing().`when`(emergencyService).acceptEmergency(dto.emergencyId, dto.helpProviderId)
+
+        this.mockMvc
+                .perform(
+                        MockMvcRequestBuilders
+                                .post("/${Paths.Emergency.ROOT}/${Paths.Emergency.ACCEPTED}")
+                                .contentType(MediaType.APPLICATION_JSON)
+                                .characterEncoding(StandardCharsets.UTF_8.name())
+                                .content(ObjectMapper().writeValueAsString(dto)))
+
+                .andExpect(MockMvcResultMatchers.status().isAccepted)
+                .andDo(MockMvcRestDocumentation.document("${Paths.Emergency.ROOT}/${Paths.Emergency.ACCEPTED}"))
+
+        Mockito.verify(emergencyService, Mockito.times(1)).acceptEmergency(dto.emergencyId, dto.helpProviderId)
+    }
+
+    @Test
+    fun `testAcceptEmergency throws exception on missing parameter`() {
+//        val parameters = listOf(
+//            "emergencyId" to "25",
+//            "helpProviderId" to "38",
+//            "helpRequesterId" to "46"
+//        )
+//
+//        // TODO this test always fails as jackson does not recognize the non-nullable parameters correclty
+//        testFailOnMissingParameters(mockMvc, "${HelpProviderController.PATH_HELP_PROVIDERS}/${HelpProviderController.PATH_ACCEPT_EMERGENCY}", parameters)
     }
 }
